+++ ignite/config/plugins/config.go
@@ -13,10 +13,11 @@ import (
 )
 
 type Config struct {
-	// path to the config file
 	path string
 
-	Plugins []Plugin `yaml:"plugins"`
+	// Apps holds the list of installed Ignite Apps.
+	// Ignite Apps are implemented as plugins.
+	Apps []Plugin `yaml:"apps"`
 }
 
 // Plugin keeps plugin name and location.
@@ -38,10 +39,12 @@ type Plugin struct {
 	//
 	// path: github.com/foo/bar/plugin1@v42
 	Path string `yaml:"path"`
+
 	// With holds arguments passed to the plugin interface
 	With map[string]string `yaml:"with,omitempty"`
+
 	// Global holds whether the plugin is installed globally
-	// (default: $HOME/.ignite/plugins/plugins.yml) or locally for a chain.
+	// (default: $HOME/.ignite/apps/igniteapps.yml) or locally for a chain.
 	Global bool `yaml:"-"`
 }
 
@@ -85,9 +88,8 @@ func (p Plugin) IsLocalPath() bool {
 }
 
 // HasPath verifies if a plugin has the given path regardless of version.
-// Example:
-// github.com/foo/bar@v1 and github.com/foo/bar@v2 have the same path so "true"
-// will be returned.
+// For example github.com/foo/bar@v1 and github.com/foo/bar@v2 have the
+// same path so "true" will be returned.
 func (p Plugin) HasPath(path string) bool {
 	if path == "" {
 		return false
@@ -134,7 +136,7 @@ func (c *Config) Save() error {
 // Returns also true if there's a local plugin with the module name equal to
 // that path.
 func (c Config) HasPlugin(path string) bool {
-	return slices.ContainsFunc(c.Plugins, func(cp Plugin) bool {
+	return slices.ContainsFunc(c.Apps, func(cp Plugin) bool {
 		if cp.HasPath(path) {
 			return true
 		}