+++ ignite/pkg/cosmosanalysis/app/app.go
@@ -7,18 +7,23 @@ import (
 	"go/format"
 	"go/parser"
 	"go/token"
+	"os"
 	"path/filepath"
 
 	"github.com/pkg/errors"
 
 	"github.com/ignite/cli/ignite/pkg/cosmosanalysis"
+	"github.com/ignite/cli/ignite/pkg/cosmosver"
 	"github.com/ignite/cli/ignite/pkg/goanalysis"
+	"github.com/ignite/cli/ignite/pkg/goenv"
+	"github.com/ignite/cli/ignite/pkg/gomodule"
 	"github.com/ignite/cli/ignite/pkg/xast"
 )
 
 const (
-	appWiringImport     = "cosmossdk.io/depinject"
-	appWiringCallMethod = "Inject"
+	appWiringImport      = "cosmossdk.io/depinject"
+	appWiringCallMethod  = "Inject"
+	registerRoutesMethod = "RegisterAPIRoutes"
 )
 
 // CheckKeeper checks for the existence of the keeper with the provided name in the app structure.
@@ -96,6 +101,10 @@ func FindRegisteredModules(chainRoot string) (modules []string, err error) {
 		return nil, err
 	}
 
+	// The modules registered by Cosmos SDK `rumtime.App` are included
+	// when the app registers API modules though the `App` instance.
+	var includeRuntimeModules bool
+
 	// Loop on package's files
 	for _, f := range appPkg.Files {
 		fileImports := goanalysis.FormatImports(f)
@@ -118,6 +127,11 @@ func FindRegisteredModules(chainRoot string) (modules []string, err error) {
 				return xast.ErrStop
 			}
 
+			// Check if Cosmos SDK runtime App is called to register API routes
+			if !includeRuntimeModules {
+				includeRuntimeModules = checkRuntimeAppCalled(n)
+			}
+
 			// Find modules in RegisterAPIRoutes declaration
 			if pkgs := findRegisterAPIRoutesRegistrations(n); pkgs != nil {
 				for _, p := range pkgs {
@@ -127,6 +141,7 @@ func FindRegisteredModules(chainRoot string) (modules []string, err error) {
 					}
 					modules = append(modules, importModule)
 				}
+
 				return xast.ErrStop
 			}
 
@@ -136,6 +151,19 @@ func FindRegisteredModules(chainRoot string) (modules []string, err error) {
 			return nil, err
 		}
 	}
+
+	// Try to find the modules registered in Cosmos SDK `runtime.App`.
+	// This is required to properly generate OpenAPI specs for these
+	// modules when `app.App.RegisterAPIRoutes` is called.
+	if includeRuntimeModules {
+		runtimeModules, err := findRuntimeRegisteredModules(chainRoot)
+		if err != nil {
+			return nil, err
+		}
+
+		modules = append(modules, runtimeModules...)
+	}
+
 	return modules, nil
 }
 
@@ -270,7 +298,7 @@ func findRegisterAPIRoutesRegistrations(n ast.Node) []string {
 		return nil
 	}
 
-	if funcLitType.Name.Name != "RegisterAPIRoutes" {
+	if funcLitType.Name.Name != registerRoutesMethod {
 		return nil
 	}
 
@@ -305,4 +333,118 @@ func findRegisterAPIRoutesRegistrations(n ast.Node) []string {
 	}
 
 	return packagesRegistered
+}
+
+func checkRuntimeAppCalled(n ast.Node) bool {
+	funcLitType, ok := n.(*ast.FuncDecl)
+	if !ok {
+		return false
+	}
+
+	if funcLitType.Name.Name != registerRoutesMethod {
+		return false
+	}
+
+	for _, stmt := range funcLitType.Body.List {
+		exprStmt, ok := stmt.(*ast.ExprStmt)
+		if !ok {
+			continue
+		}
+
+		exprCall, ok := exprStmt.X.(*ast.CallExpr)
+		if !ok {
+			continue
+		}
+
+		exprFun, ok := exprCall.Fun.(*ast.SelectorExpr)
+		if !ok || exprFun.Sel.Name != registerRoutesMethod {
+			continue
+		}
+
+		exprSel, ok := exprFun.X.(*ast.SelectorExpr)
+		if !ok || exprSel.Sel.Name != "App" {
+			continue
+		}
+
+		identType, ok := exprSel.X.(*ast.Ident)
+		if !ok || identType.Name != "app" {
+			continue
+		}
+
+		return true
+	}
+
+	return false
+}
+
+func findRuntimeRegisteredModules(chainRoot string) ([]string, error) {
+	// Resolve the absolute path to the Cosmos SDK module
+	cosmosPath, err := resolveCosmosPackagePath(chainRoot)
+	if err != nil {
+		return nil, err
+	}
+
+	var modules []string
+
+	// When runtime package doesn't exists it means is an older Cosmos SDK version,
+	// so all the module API registrations are defined within user's app.
+	path := filepath.Join(cosmosPath, "runtime", "app.go")
+	if _, err := os.Stat(path); os.IsNotExist(err) {
+		return modules, nil
+	}
+
+	f, _, err := xast.ParseFile(path)
+	if err != nil {
+		return nil, err
+	}
+
+	imports := goanalysis.FormatImports(f)
+	err = xast.Inspect(f, func(n ast.Node) error {
+		if pkgs := findRegisterAPIRoutesRegistrations(n); pkgs != nil {
+			for _, p := range pkgs {
+				if m := imports[p]; m != "" {
+					modules = append(modules, m)
+				}
+			}
+			return xast.ErrStop
+		}
+		return nil
+	})
+
+	if err != nil {
+		return nil, err
+	}
+	return modules, nil
+}
+
+func resolveCosmosPackagePath(chainRoot string) (string, error) {
+	modFile, err := gomodule.ParseAt(chainRoot)
+	if err != nil {
+		return "", err
+	}
+
+	deps, err := gomodule.ResolveDependencies(modFile)
+	if err != nil {
+		return "", err
+	}
+
+	var pkg string
+	for _, dep := range deps {
+		if dep.Path == cosmosver.CosmosModulePath {
+			pkg = dep.String()
+			break
+		}
+	}
+
+	if pkg == "" {
+		return "", errors.New("Cosmos SDK package version not found")
+	}
+
+	// Check path of the package directory within Go's module cache
+	path := filepath.Join(goenv.GoModCache(), pkg)
+	info, err := os.Stat(path)
+	if os.IsNotExist(err) || !info.IsDir() {
+		return "", errors.New("local path to Cosmos SDK package not found")
+	}
+	return path, nil
 }